refactor: extract diff parsing util
