+++ apps/api/src/handlers/diff.ts
@@ -1,6 +1,7 @@
 import { DiffFile, DiffMode } from "@diffview/shared";
 
-import { getDiffFiles, parseDiffFiles, getDiffPatches, splitPatchByFile } from "@/utils/git";
+import { parseDiffFiles } from "@/utils/diff-files";
+import { getDiffFiles, getDiffPatches, splitPatchByFile } from "@/utils/git";
 import { json } from "@/utils/http";
 
 export default async function handleDiff(request: Request): Promise<Response> {
